# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ADIRRA API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/glossary/terms/{term_id}": {
      "get": {
        "tags": [
          "glossary"
        ],
        "summary": "Get Term",
        "operationId": "get_term_glossary_terms__term_id__get",
        "parameters": [
          {
            "name": "term_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Term Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "glossary"
        ],
        "summary": "Delete Term",
        "operationId": "delete_term_glossary_terms__term_id__delete",
        "parameters": [
          {
            "name": "term_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Term Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/glossary/terms": {
      "put": {
        "tags": [
          "glossary"
        ],
        "summary": "Create Or Update Term",
        "operationId": "create_or_update_term_glossary_terms_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlossaryTermSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/glossary/terms/{term_id}/ai-suggest": {
      "post": {
        "tags": [
          "glossary"
        ],
        "summary": "Ai Suggest",
        "operationId": "ai_suggest_glossary_terms__term_id__ai_suggest_post",
        "parameters": [
          {
            "name": "term_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Term Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/glossary/terms/{term_id}/confirm": {
      "post": {
        "tags": [
          "glossary"
        ],
        "summary": "Confirm Term",
        "description": "Confirm a glossary term (Steward review, Phase E2b).\n\nSets the term status to 'approved' \u2014 this confirms the term for EVERY column\nit is linked to (glossary status is a term-level property, not per-link).",
        "operationId": "confirm_term_glossary_terms__term_id__confirm_post",
        "parameters": [
          {
            "name": "term_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Term Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TermDecisionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/glossary/terms/{term_id}/reject": {
      "post": {
        "tags": [
          "glossary"
        ],
        "summary": "Reject Term",
        "description": "Reject a glossary term \u2014 reverts status to 'draft' for re-editing (Phase E2b).",
        "operationId": "reject_term_glossary_terms__term_id__reject_post",
        "parameters": [
          {
            "name": "term_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Term Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TermDecisionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/glossary/terms/{term_id}/ai-suggest-fields": {
      "post": {
        "tags": [
          "glossary"
        ],
        "summary": "Ai Suggest Fields",
        "description": "Generate AI suggestions for specific fields of a term.\n\nBody: { \"fields\": [\"business_description\", \"detailed_description\", ...] }\nReturns the same shape as ai-suggest but only the requested fields are generated.",
        "operationId": "ai_suggest_fields_glossary_terms__term_id__ai_suggest_fields_post",
        "parameters": [
          {
            "name": "term_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Term Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/glossary/terms/{term_id}/crr-context": {
      "post": {
        "tags": [
          "glossary"
        ],
        "summary": "Generate Crr Context",
        "description": "Generate CRR3 regulatory context for a term.",
        "operationId": "generate_crr_context_glossary_terms__term_id__crr_context_post",
        "parameters": [
          {
            "name": "term_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Term Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/glossary/terms/{term_id}/dpm-context": {
      "post": {
        "tags": [
          "glossary"
        ],
        "summary": "Generate Dpm Context",
        "description": "Generate DPM 2.0 reporting context for a term.",
        "operationId": "generate_dpm_context_glossary_terms__term_id__dpm_context_post",
        "parameters": [
          {
            "name": "term_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Term Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/glossary/cross-ref": {
      "get": {
        "tags": [
          "glossary"
        ],
        "summary": "Cross Ref",
        "description": "Find glossary terms whose related_objects include the given catalog reference.",
        "operationId": "cross_ref_glossary_cross_ref_get",
        "parameters": [
          {
            "name": "ref",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ref"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/settings/import/glossary": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Import Glossary",
        "operationId": "import_glossary_settings_import_glossary_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_glossary_settings_import_glossary_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_glossary_settings_import_glossary_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "merge_mode": {
            "type": "string",
            "title": "Merge Mode",
            "default": "skip"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_glossary_settings_import_glossary_post"
      },
...
      "GlossaryTermSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": ""
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "default": ""
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "business_description": {
            "type": "string",
            "title": "Business Description",
            "default": ""
          },
          "detailed_description": {
            "type": "string",
            "title": "Detailed Description",
            "default": ""
          },
          "synonyms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Synonyms",
            "default": []
          },
          "related_objects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Related Objects",
            "default": []
          },
          "steward": {
            "type": "string",
            "title": "Steward",
            "default": ""
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "draft"
          },
          "CRR_context": {
            "type": "string",
            "title": "Crr Context",
            "default": ""
          },
          "DPM_context": {
            "type": "string",
            "title": "Dpm Context",
            "default": ""
          },
          "ai_generated_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ai Generated Fields",
            "default": []
          },
          "last_updated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated"
          },
          "last_reviewed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reviewed"
          }
        },
        "type": "object",
        "title": "GlossaryTermSchema"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TermDecisionRequest": {
        "properties": {
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "decided_by_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By Role"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "TermDecisionRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}